Network agent + event-messaging  system
Create an Event manager that will allow objects to trigger and
listen for events.

Create a Network Agent that will act as a wrapper
for the underlying websocket library, and conneccted
it to the messaging system.

Added a test to use the event-messaging system to
connect/disconnect to the server, and display a message
from the server.